requirements.txt
bcrypt>=4.3.0
fastapi>=0.120.0
loguru>=0.7.3
passlib[bcrypt]>=1.7.4
pydantic>=2.11.9
pyjwt>=2.10.1
pymongo>=4.15.3
python-multipart>=0.0.20
requests>=2.32.5
uvicorn[standard]>=0.38.0
python-dotenv>=1.2.1
python-jose[cryptography]>=3.5.0
